Compute smooth Mandelbrot iteration counts
Store Mandelbrot iteration counts as floats and compute continuous escape
values in the CPU reference path and compute shader.  Switch the GPU
iteration texture to R32_FLOAT and interpolate adjacent palette entries
when mapping fractional counts to color.

+++ libs/core/include/core/MandelCpu.h
@@ -6,7 +6,7 @@
 namespace core
 {
 
-int mandel_cpu_pixel(const MandelParams &params, int px, int py);
+float mandel_cpu_pixel(const MandelParams &params, int px, int py);
 MandelIterationBuffer render_mandel_cpu(const MandelParams &params);
 
 } // namespace core
